feat: pre check
