+++ packages/ai-workspace-common/src/components/canvas/top-toolbar/publish-template-button.tsx
@@ -1,5 +1,5 @@
 import React, { useMemo, useCallback, useState } from 'react';
-import { Button, Tooltip } from 'antd';
+import { Button, Tooltip, message } from 'antd';
 import { useTranslation } from 'react-i18next';
 import { CreateWorkflowAppModal } from '@refly-packages/ai-workspace-common/components/workflow-app/create-modal';
 import { useListWorkflowApps } from '@refly-packages/ai-workspace-common/queries';
@@ -10,6 +10,8 @@ import { useSkillResponseLoadingStatus } from '@refly-packages/ai-workspace-comm
 import { TurnRight } from 'refly-icons';
 import { cn } from '@refly/utils/cn';
 import { PublishTemplatePopover } from './publish-template-popover';
+import { useRealtimeCanvasData } from '@refly-packages/ai-workspace-common/hooks/canvas/use-realtime-canvas-data';
+import { useVariablesManagement } from '@refly-packages/ai-workspace-common/hooks/use-variables-management';
 
 interface PublishTemplateButtonProps {
   canvasId: string;
@@ -37,17 +39,128 @@ const PublishTemplateButton = React.memo(
       return result;
     }, [workflowAppsData, canvasId]);
 
+    // Get canvas data and variables for validation
+    const { data: workflowVariables } = useVariablesManagement(canvasId);
+    const { nodes } = useRealtimeCanvasData();
+
+    // Filter skillResponse nodes for validation
+    const skillResponseNodesForValidation = useMemo(() => {
+      return nodes.filter((node) => node.type === 'skillResponse');
+    }, [nodes]);
+
+    // Canvas validation checks
+    const canvasValidation = useMemo(() => {
+      // Check if there are no user input variables
+      const hasNoVariables = !workflowVariables || workflowVariables.length === 0;
+
+      // Check for failed or unrun skillResponse nodes
+      // Returns true if any skillResponse node has status 'failed', 'init', or undefined
+      const hasFailedOrUnrunNodes = skillResponseNodesForValidation.some((node) => {
+        const status = (node.data?.metadata as any)?.status;
+        return status === 'failed' || status === 'init' || !status;
+      });
+
+      // Check if there are no skillResponse nodes
+      // Returns true if there are no Agent nodes in the canvas
+      const hasNoSkillResponseNodes = skillResponseNodesForValidation.length === 0;
+
+      return {
+        hasNoVariables,
+        hasFailedOrUnrunNodes,
+        hasNoSkillResponseNodes,
+      };
+    }, [workflowVariables, skillResponseNodesForValidation]);
+
+    // Validate canvas before publishing
+    // Returns true if validation passes, false otherwise
+    const validateCanvas = useCallback(() => {
+      // Priority: User Input > No Agents > Agent Not Run or Failed
+      if (canvasValidation.hasNoVariables) {
+        // Show toast
+        message.error(t('workflowApp.validationNoUserInputs'));
+        // Trigger canvas fitView event
+        window.dispatchEvent(
+          new CustomEvent('refly:canvas:fitView', {
+            detail: { canvasId },
+          }),
+        );
+        return false;
+      }
+
+      if (canvasValidation.hasNoSkillResponseNodes) {
+        // Show toast
+        message.error(t('workflowApp.validationNoAgents'));
+        // Trigger canvas fitView event
+        window.dispatchEvent(
+          new CustomEvent('refly:canvas:fitView', {
+            detail: { canvasId },
+          }),
+        );
+        return false;
+      }
+
+      if (canvasValidation.hasFailedOrUnrunNodes) {
+        // Get failed or unrun node IDs for highlighting
+        const failedOrUnrunNodeIds = skillResponseNodesForValidation
+          .filter((node) => {
+            if (!node?.id) return false;
+            const status = (node.data?.metadata as any)?.status;
+            return status === 'failed' || status === 'init' || !status;
+          })
+          .map((node) => node.id)
+          .filter(Boolean) as string[];
+
+        // Show toast
+        message.error(t('workflowApp.validationAgentsNotRun'));
+
+        // Only trigger highlight if we have valid node IDs
+        if (failedOrUnrunNodeIds.length > 0) {
+          // Trigger canvas fitView and highlight event
+          window.dispatchEvent(
+            new CustomEvent('refly:canvas:fitViewAndHighlight', {
+              detail: {
+                canvasId,
+                nodeIds: failedOrUnrunNodeIds,
+              },
+            }),
+          );
+        } else {
+          // Fallback: just fit view if no valid node IDs found
+          window.dispatchEvent(
+            new CustomEvent('refly:canvas:fitView', {
+              detail: { canvasId },
+            }),
+          );
+        }
+        return false;
+      }
+
+      return true;
+    }, [canvasValidation, canvasId, skillResponseNodesForValidation, t]);
+
     const handlePublishToCommunity = useCallback(async () => {
       // Make sure the canvas data is synced to the remote
       await forceSyncState({ syncRemote: true });
+
+      // Validate canvas before opening modal
+      if (!validateCanvas()) {
+        return;
+      }
+
       setCreateTemplateModalVisible(true);
-    }, [forceSyncState]);
+    }, [forceSyncState, validateCanvas]);
 
     const handleUpdateTemplate = useCallback(async () => {
       // Make sure the canvas data is synced to the remote
       await forceSyncState({ syncRemote: true });
+
+      // Validate canvas before opening modal
+      if (!validateCanvas()) {
+        return;
+      }
+
       setCreateTemplateModalVisible(true);
-    }, [forceSyncState]);
+    }, [forceSyncState, validateCanvas]);
 
     const handlePublishSuccess = useCallback(async () => {
       // Refresh workflow apps data after successful publish
